[FIX] sale: recompute taxes on company change
Since e6415be5ce6077702836779cf784e16f7e72a9c7, it is
forbidden to use taxes on Sales Order Lines from another company
than the Sales Order one (it was breaking later in the flows anyway).

Nevertheless, in some flows, modifying the company of an existing SO
can make sense, but you wouldn't be able to save the change if taxes
were set on the lines (unless you trigger the 'Update Taxes' button).

This commit will make sure that taxes are automatically recomputed
on company change, s.t. the user doesn't face multi-company errors.
It'll magically change taxes & amounts, but we consider that it's
an expected behavior when modifying the order company (and an
advanced flow for which users must be aware of the consequences).

Anyway, the user will always be able to discard the changes (company
& company-based recomputations) if needed.

opw-3234905

closes odoo/odoo#120774

X-original-commit: e2e97ae66604a441e91e0efedd8d3b10c4219d24
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>